scoreextractor moved to cmd
