+++ cmd/scoreextractor/checkpoints_test_data/1615018620.png
