+++ cmd/scoreextractor/checkpoints_test_data/1615022713.png
