+++ cmd/scoreextractor/checkpoints_test_data/1615041930.png
