+++ cmd/scoreextractor/checkpoints_test_data/1615043111.ram
