+++ cmd/scoreextractor/checkpoints_test_data/1615047878.ram
