+++ cmd/scoreextractor/checkpoints_test_data/1615048176.png
