+++ cmd/scoreextractor/score_extractor.go
