+++ cmd/scoreextractor/score_extractor_test.go
