+++ nes/ui/gameview.go
@@ -13,7 +13,7 @@ import (
 	"github.com/go-gl/gl/v2.1/gl"
 	"github.com/go-gl/glfw/v3.2/glfw"
 	log "github.com/sirupsen/logrus"
-	extractor "github.com/skycoin/cx-aigym-nes/extractor/scoreextractor"
+	extractor "github.com/skycoin/cx-aigym-nes/cmd/scoreextractor"
 	"github.com/skycoin/cx-aigym-nes/nes/nes"
 )
